refactor file hierarchy
