+++ src/console/console.rs
@@ -1,0 +1,1 @@
+/* do something. */
