+++ src/driver.rs
@@ -1,0 +1,2 @@
+pub mod ps2;
+pub mod vga;
