+++ src/driver/ps2/keyboard.rs
@@ -1,5 +1,5 @@
 use super::control::{test_status_now, Status};
-use crate::pmio::Port;
+use crate::raw_io::pmio::Port;
 
 static KEYBOARD_PORT: Port = Port::new(0x60);
 
@@ -139,6 +139,9 @@ pub fn get_key_event() -> Option<KeyEvent> {
 	Some(scancode_to_keyevent(page, raw_scancode))
 }
 
+/*
+ * Generated by script. do not touch
+*/
 static SCAN_CODE_SET1: [[Key; 128]; 2] = [
 	[
 		Key::Unused,
