+++ src/driver/vga.rs
@@ -1,0 +1,1 @@
+pub mod text_vga;
